# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server Market API.",
        "title": "Market API",
        "contact": {
            "name": "Seyma TUTAN GUN",
            "email": "seymatutan@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:9000",
    "basePath": "/",
    "paths": {
...
        "/api/payment/complete": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Complete payment with qr code and total amount",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payment"
                ],
                "summary": "Complete Payment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "write Bearer before token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "req params",
                        "name": "PaymentReq",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.PaymentReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/api/wallet/get": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get Wallet Details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wallet"
                ],
                "summary": "Get Wallet Details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "write Bearer before token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.WalletRes"
                        }
                    }
                }
            }
        },
        "/api/wallet/load": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Load Money to Wallet",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wallet"
                ],
                "summary": "Load Money to Wallet",
                "parameters": [
                    {
                        "type": "string",
                        "description": "write Bearer before token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "req params",
                        "name": "WalletTransactionReqBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.WalletTransactionReqBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/wallet/qr": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Generate Login User QRCode for payment.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wallet"
                ],
                "summary": "Generate QRCode for payment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "write Bearer before token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.QRCodeRes"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.PaymentReq": {
            "type": "object",
            "required": [
                "amount",
                "qr_code"
            ],
            "properties": {
                "amount": {
                    "type": "number"
                },
                "qr_code": {
                    "type": "string"
                }
            }
        },
        "model.QRCodeRes": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                }
            }
        },
...
        "model.WalletRes": {
            "type": "object",
            "properties": {
                "current_amount": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "unique_key": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "model.WalletTransactionReqBody": {
            "type": "object",
            "required": [
                "amount",
                "card_number",
                "cvc",
                "transaction_type",
                "wallet_id"
            ],
            "properties": {
                "amount": {
                    "type": "number"
                },
                "card_number": {
                    "type": "string"
                },
                "cvc": {
                    "type": "integer"
                },
                "transaction_type": {
                    "type": "integer"
                },
                "wallet_id": {
                    "type": "integer"
                }
            }
        }
    }
}